refactor: clean duplicated code in tests
