+++ crates/runtime/tests/allow_codegen.rs
@@ -1,19 +1,11 @@
-use lagon_runtime::{options::RuntimeOptions, Runtime};
 use lagon_runtime_http::{Request, Response, RunResult};
 use lagon_runtime_isolate::{options::IsolateOptions, Isolate};
-use std::sync::Once;
 
-fn setup() {
-    static START: Once = Once::new();
-
-    START.call_once(|| {
-        Runtime::new(RuntimeOptions::default().allow_code_generation(true));
-    });
-}
+mod utils;
 
 #[tokio::test]
 async fn allow_eval() {
-    setup();
+    utils::setup_allow_codegen();
     let mut isolate = Isolate::new(IsolateOptions::new(
         "export function handler() {
     const result = eval('1 + 1')
@@ -32,7 +24,7 @@ async fn allow_eval() {
 
 #[tokio::test]
 async fn allow_function() {
-    setup();
+    utils::setup_allow_codegen();
     let mut isolate = Isolate::new(IsolateOptions::new(
         "export function handler() {
     const result = new Function('return 1 + 1')
