+++ crates/serverless/tests/deployments.rs
@@ -1,6 +1,5 @@
 use anyhow::Result;
 use dashmap::DashMap;
-use lagon_runtime::{options::RuntimeOptions, Runtime};
 use lagon_runtime_utils::Deployment;
 use lagon_serverless::{
     cronjob::Cronjob, deployments::downloader::FakeDownloader, serverless::start,
@@ -8,24 +7,16 @@ use lagon_serverless::{
 use serial_test::serial;
 use std::{
     collections::{HashMap, HashSet},
-    sync::{Arc, Once},
+    sync::Arc,
 };
 use tokio::sync::Mutex;
 
-fn setup() {
-    static START: Once = Once::new();
-
-    START.call_once(|| {
-        dotenv::dotenv().expect("Failed to load .env file");
-
-        Runtime::new(RuntimeOptions::default());
-    });
-}
+mod utils;
 
 #[tokio::test]
 #[serial]
 async fn simple() -> Result<()> {
-    setup();
+    utils::setup();
     let deployments = Arc::new(DashMap::new());
     deployments.insert(
         "127.0.0.1:4000".into(),
@@ -62,7 +53,7 @@ async fn simple() -> Result<()> {
 #[tokio::test]
 #[serial]
 async fn reuse_isolate() -> Result<()> {
-    setup();
+    utils::setup();
     let deployments = Arc::new(DashMap::new());
     deployments.insert(
         "127.0.0.1:4001".into(),
@@ -103,7 +94,7 @@ async fn reuse_isolate() -> Result<()> {
 #[tokio::test]
 #[serial]
 async fn reuse_isolate_across_domains() -> Result<()> {
-    setup();
+    utils::setup();
     let deployments = Arc::new(DashMap::new());
     let deployment = Arc::new(Deployment {
         id: "counter".into(),
