+++ crates/serverless/tests/requests.rs
@@ -2,7 +2,6 @@ use anyhow::Result;
 use dashmap::DashMap;
 use futures::StreamExt;
 use hyper::body::Bytes;
-use lagon_runtime::{options::RuntimeOptions, Runtime};
 use lagon_runtime_utils::Deployment;
 use lagon_serverless::{
     cronjob::Cronjob, deployments::downloader::FakeDownloader, serverless::start,
@@ -10,24 +9,16 @@ use lagon_serverless::{
 use serial_test::serial;
 use std::{
     collections::{HashMap, HashSet},
-    sync::{Arc, Once},
+    sync::Arc,
 };
 use tokio::sync::Mutex;
 
-fn setup() {
-    static START: Once = Once::new();
-
-    START.call_once(|| {
-        dotenv::dotenv().expect("Failed to load .env file");
-
-        Runtime::new(RuntimeOptions::default());
-    });
-}
+mod utils;
 
 #[tokio::test]
 #[serial]
 async fn returns_correct_http() -> Result<()> {
-    setup();
+    utils::setup();
     let deployments = Arc::new(DashMap::new());
     deployments.insert(
         "127.0.0.1:4000".into(),
@@ -67,7 +58,7 @@ async fn returns_correct_http() -> Result<()> {
 #[tokio::test]
 #[serial]
 async fn returns_correct_path() -> Result<()> {
-    setup();
+    utils::setup();
     let deployments = Arc::new(DashMap::new());
     deployments.insert(
         "127.0.0.1:4000".into(),
@@ -115,7 +106,7 @@ async fn returns_correct_path() -> Result<()> {
 #[tokio::test]
 #[serial]
 async fn forwards_headers() -> Result<()> {
-    setup();
+    utils::setup();
     let deployments = Arc::new(DashMap::new());
     deployments.insert(
         "127.0.0.1:4000".into(),
@@ -154,7 +145,7 @@ async fn forwards_headers() -> Result<()> {
 #[tokio::test]
 #[serial]
 async fn stream_sequentially() -> Result<()> {
-    setup();
+    utils::setup();
     let deployments = Arc::new(DashMap::new());
     deployments.insert(
         "127.0.0.1:4000".into(),
